fix a couple clippy warning.

+++ unsafe_collection/src/no_hash.rs
@@ -38,7 +38,7 @@ impl Hasher for NoHasher {
 
     #[inline]
     fn write_u64(&mut self, i: u64) {
-        self.0 = i as u64
+        self.0 = i
     }
 
     #[inline]
